apply per-invoice escrow hold period and tighten processor interfaces
store and enforce invoice escrowHoldPeriod during payment/release scheduling; add invalid oracle answer guard; migrate storage config numeric fields to uint96 with deploy/setup casting; replace custom IERC20 usage with safe decimals lookup fallback; align interfaces/docs/tests for UnsupportedToken and hold-period behavior

+++ test/utils/BaseSetUp.sol
@@ -38,9 +38,9 @@ abstract contract BaseSetUp is Test {
             owner: admin,
             feeReceiver: feeReceiver,
             marketplace: address(this),
-            feeRate: FEE_RATE,
-            defaultHoldPeriod: DEFAULT_HOLD_PERIOD,
-            gasThreshold: DEFAULT_GAS_Threshold
+            feeRate: uint96(FEE_RATE),
+            defaultHoldPeriod: uint96(DEFAULT_HOLD_PERIOD),
+            gasThreshold: uint96(DEFAULT_GAS_Threshold)
         });
 
         vm.prank(admin);
